Move `distributeSurplus` to new `Balance.Distribute` module
